server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
ultralytics>=8.0.0
torch>=2.0.0
python-multipart>=0.0.6
websockets>=11.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
pillow>=10.0.0
